refactor(dashboard): extract utility functions for data fetching and response parsing
- Introduce `parseRows` and `buildPgrestFetch` utility functions to streamline data fetching and response handling in Dashlet and DashletSettings.
- Replace inline logic with these utility functions to enhance code readability and maintainability, improving overall structure and organization of the data fetching process.
